visualization/polus-graph-pyramid-builder-plugin/src/requirements.txt
pandas==1.3.5
matplotlib==3.5.1
numpy==1.21.4
Pillow==8.4.0
imageio==2.5.0
scipy==1.7.3